requirements.txt
asgiref==3.4.1
Django==3.2.8
Pillow==8.4.0
python-slugify==5.0.2
pytz==2021.3
sqlparse==0.4.2
text-unidecode==1.3
